backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pika==1.3.2
pymysql==1.1.1
python-dotenv==1.2.2
bcrypt==4.2.0
PyJWT==2.9.0
cryptography==42.0.5
requests==2.32.3
